refactor: extract components and simplify App
