+++ src/ui/components/Footer.tsx
@@ -1,0 +1,87 @@
+import { Github, Linkedin, Mail, MapPin, Phone, Twitter, Zap } from "lucide-react";
+
+export default function Footer() {
+  return (
+    <footer className="bg-gray-900 text-white">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-16">
+        <div className="grid md:grid-cols-4 gap-8">
+          <div className="space-y-4">
+            <div className="flex items-center space-x-2">
+              <div className="h-8 w-8 bg-gradient-to-r from-blue-600 to-purple-600 rounded-lg flex items-center justify-center">
+                <Zap className="h-5 w-5 text-white" />
+              </div>
+              <span className="text-xl font-bold">DataFlow</span>
+            </div>
+            <p className="text-gray-400 leading-relaxed">
+              Empowering businesses with AI-powered analytics and insights that drive growth and innovation.
+            </p>
+            <div className="flex space-x-4">
+              {[Twitter, Linkedin, Github].map((Icon, index) => (
+                <button
+                  key={index}
+                  className="p-2 bg-gray-800 rounded-lg hover:bg-gray-700 transition-colors duration-200 transform hover:scale-110"
+                >
+                  <Icon className="h-5 w-5" />
+                </button>
+              ))}
+            </div>
+          </div>
+          <div>
+            <h4 className="font-semibold mb-4">Product</h4>
+            <ul className="space-y-2">
+              {["Features", "Pricing", "API Documentation", "Integrations"].map((item) => (
+                <li key={item}>
+                  <button className="text-gray-400 hover:text-white dark:text-gray-300 dark:hover:text-gray-100 transition-colors duration-200">
+                    {item}
+                  </button>
+                </li>
+              ))}
+            </ul>
+          </div>
+          <div>
+            <h4 className="font-semibold mb-4">Company</h4>
+            <ul className="space-y-2">
+              {["About Us", "Careers", "Blog", "Press"].map((item) => (
+                <li key={item}>
+                  <button className="text-gray-400 hover:text-white dark:text-gray-300 dark:hover:text-gray-100 transition-colors duration-200">
+                    {item}
+                  </button>
+                </li>
+              ))}
+            </ul>
+          </div>
+          <div>
+            <h4 className="font-semibold mb-4">Contact</h4>
+            <div className="space-y-3">
+              <div className="flex items-center space-x-3 text-gray-400">
+                <Mail className="h-5 w-5" />
+                <span>hello@dataflow.com</span>
+              </div>
+              <div className="flex items-center space-x-3 text-gray-400">
+                <Phone className="h-5 w-5" />
+                <span>+1 (555) 123-4567</span>
+              </div>
+              <div className="flex items-center space-x-3 text-gray-400">
+                <MapPin className="h-5 w-5" />
+                <span>San Francisco, CA</span>
+              </div>
+            </div>
+          </div>
+        </div>
+        <div className="border-t border-gray-800 dark:border-gray-700 mt-12 pt-8 flex flex-col md:flex-row justify-between items-center">
+          <p className="text-gray-400 dark:text-gray-300 text-sm">© {new Date().getFullYear()} DataFlow. All rights reserved.</p>
+          <div className="flex space-x-6 mt-4 md:mt-0">
+            {["Privacy Policy", "Terms of Service", "Cookie Policy"].map((item) => (
+              <button
+                key={item}
+                className="text-gray-400 hover:text-white dark:text-gray-300 dark:hover:text-gray-100 text-sm transition-colors duration-200"
+              >
+                {item}
+              </button>
+            ))}
+          </div>
+        </div>
+      </div>
+    </footer>
+  );
+}
